requirements.txt
Django==1.9.6
django-countries==3.4.1
djangorestframework==3.3.3
django-crispy-forms==1.6.0
django-filter==0.13.0
django_extensions==1.6.7
gunicorn==19.5.0
dj-database-url==0.4.1
psycopg2==2.6.1
python-dateutil==2.6.0
whitenoise==3.1

# Tests
Faker==0.7.7
factory-boy==2.8.1
pytest==2.9.1
pytest-django==2.9.1
pytest-sugar==0.7.1
raven==5.19.0
freezegun==0.3.8
